requirements.txt
tensorflow
matplotlib
tqdm
pillow
numpy
pandas